Do not report false function point info
In [dc278a6] basic user feedback was added to inform the user when not
within a function point.  When testing I'm guessing I tested the "not in
function point" path first and after that the "in function point" case.
I wrongfully assumed everything was working ok and that the "not in
function point" message that I saw in the second case was just the
previous one (which, I know, was ridiculously wrong).

By returning from the function instead of breaking the for loop (see
code change) we prevent the warning message from appearing to the user
when everything is in fact according to plan.

Closes #23