POC: Load queries from `.scm` files
"This is the way": We will load relevant SCM files from file system,
based on what we need and the programming language of the current
buffer.  (which is still too naive in case of PHP/HTML mixtures, but
that's a problem to tackle later on).

General idea is to have TS queries for functions, conditionals and
iterations and then parse and calculate accordingly.  This will keep the
base Kokwame code a lot cleaner because it pushes the abstraction per
language down to the TS queries.