Namespace query files
Auery files loading goes down the runtime path, so if a Kokwame query
file should accidentally have the same name as an existing one, it'll
override the existing one, and we don't want that.
The query files that exist now, do not have the same name as existing
ones, but I think it's better to be safe than sorry.